relocate min_max_scale function to utils
