+++ assume/reinforcement_learning/learning_utils.py
@@ -104,21 +104,6 @@ def polyak_update(params, target_params, tau: float):
             th.add(target_param.data, param.data, alpha=tau, out=target_param.data)
 
 
-def min_max_scale(x, min_val: float, max_val: float):
-    """
-    Min-Max scaling of a value x to the range [0, 1]
-
-    Args:
-        x: value(s) to scale
-        min_val: minimum value of the parameter
-        max_val: maximum value of the parameter
-    """
-    # Avoid division by zero
-    if min_val == max_val:
-        return x
-    return (x - min_val) / (max_val - min_val)
-
-
 def linear_schedule_func(
     start: float, end: float = 0, end_fraction: float = 1
 ) -> Schedule:
